requirements.txt
colorlog
python-magic
libmagic
python-magic-bin
argparse